IAM: keep LegacyUserTeamsSearchClient separate from LegacyTeamSearchClient
Revert the previous consolidation: the user-teams subresource is wired to
its own LegacyUserTeamsSearchClient so the two search clients stay focused
on a single (resource × filter) pair. Adds unit tests for the standalone
client (pagination, validation, error propagation) and a parity subtest in
the integration suite that asserts permission strings round-trip
identically (admin and member) across dual-writer Mode 0/1/5.

Also drops the unused common.PaginationFromQuery helper that became dead
code when rest_user_team.go reverted to inline pagination.
